services/lab-workspace/docker-res/duplicated-resources/ml-lab-py/requirements.txt
# Swagger Client
certifi >= 14.05.14
six >= 1.10
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.15.1

# Environment
tensorboardX >= 1.2
click
requests
psutil
GitPython
setuptools
tqdm
future
requests-toolbelt